Section 2.
1.  example of main.ts sourcing a defined component.
<courses></courses>

2.  the courses is define under courses.component.ts as what is defined to inject.

it has a template, direcective, and even a service where it gets data.

3.  the servcies is defined for courses as courses.service.ts.

This service mocks what would be returned from an outside data source (database, restful api, etc)

Also note the directive under the courses.component uses dependency injection to set the data for 'courses;'

constructor(courseService: CourseService)  { ... }

this allows for unit testing and de coupling of the comonent to the data.

AngularJs will inject the dependencies here .